requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.35
python-barcode==0.15.1
Pillow>=10.0.0
reportlab==4.0.7
requests==2.31.0
